Related to issue 35015: Code review improvements
As Cost Adjustment Header's organization already is saved with the "legal"
organization, is not needed search it again to set as org of it lines, when
them are created. So the Cost Adjustment Line's organization will be taken
from the header's organization.

Also, as the header and lines will have the same organization is not needed to
show this field in the lines. So, the Organization field in the Cost Adjustment
Line tab has been hidden again.