Fixed issue 35064: A deactivated organization can not be activated again.
When an organization was disabled you could not activate it again as the row was not ediable. This
problem was introduced in a fix related with the code of addWritableAttribute() method of
DefaultJsonDataService class. Before this fix, It was managing the reported case properly (Activate
again a deactivated organization) because this method was handling the particular case of Organizations.

Now is taking into account this particular case (deactivate organizations) in order to manages this
particular case properly. Now it is possible to activate again a deactivate organization.