related to bug 35057: keep check invariant in rollback
After the changes introduced on issue #34816, SessionHandler.getInstance() does not begin a session when the handler was not already initilized. This causes that the checkInvariant throws an exception when calling rollback() after commitAndClose(), because the session is null when calling rollback().

The problem here was that OBBaseTest was doing that,i.e., invoking rollback inside a catch block when commitAndClose() fails. This is not needed, because when commitAndClose()  fails the rollback of the transaction is done in that very same method.

For this reason, we are keeping the checkInvariant invocation in the rollback method, and we are deleting the usage of rollback() in the catch block in OBBaseTest.