Fixed 35088. Make 2 junit tests more robust.
Both junit tests assigned that the c_country with _id 106 has the name
"Spain". However in some projects that value is directly modified/translated.
Make the test more robust to search for the country via isocode 'ES' which
is more stable. And in the json test verify the expected row via its id 106
instead of its identifier 'Spain'.