Fixed issue 35055: It was not possible to use a button ref. with list subref.
If a process button had a button reference and a list sub reference was asigned, the behavior was not correct. It was acting like the main reference was button list instead of button.
The problem was that the parameter buttons was always set, this parameter is the one which contains the labels of the buttons. If this paramater is set, the code chooses a flow in which shows those buttons and the behaviour is just like having a button list reference with a list subreference.
To fix this, the button parameter must not be set. This is the action that has been taken. It is important to know that this will not break the behavior of the button list references with list subreferences, because that case is covered by the "ob-pick-and-execute-view-window.js.ftl" In this ftl, the buttons property is set.
Now, the buttons property is set if the process ui definition is Manual, even if there are no known cases, as we supported it before, now it have to be also.

--HG--
extra : rebase_source : f352c7b5091ba669fe36867bd7232bcd0fce6c70