Related with issue34824: Mandatory columns cannot be empty nor null
It was also possible to save a record in the "GL Items" section of the "Add Payment" parameter window when a mandatory column had a null value.
This has been fixed by adding a new condition which now checks if the mandatory column has an undefined value or a null value. And if it is true, the record is not saved.

--HG--
extra : rebase_source : 1fa5e5321521fec7a164cf56a561743aeb7ecc6c