Fixed issue 35055: Buttons with list reference where not working fine.
When having a button with list reference and a process definition with a button list parameter, it was not taking the label values from the column, it was only taking them from the parameters.
The code which was taking the labels of the parameters has been removed because the way of obtaining the parameters' lables is covered in the "ob-pick-and-execute-view-window.js.ftl".
The problem with the removed piece of code was that, at first, all the parameters were processed and this caused having the "labelValue" variable filled with the label values of the parameters. As the "labelValue" variable was not empty, the code which handled the label values of the button was never reached, because it was inside an if (labelValues != null).

--HG--
extra : rebase_source : c1f8f9aecbe807b21bbb3f287c8d2f980bb75fab