fixes issue 35193: Error fired when trying to print and attach a document
With the changes introduced with the read-only pool project, the reporting engine is making use of the DalConnectionProvider to get information from the database. On the other hand, the fix for issue #32576 closes the Dal connection to avoid having connection leaks when using Dal from a subreport. Before having the read-only project that fix did not affect to the reporting engine because the PrintController class itself was the connection provider and its connection was not being closed by the subreports using Dal.

But now the same connection provider is used by the PrintController and also by the sub-reports (when they use Dal). For this reason we need to check if the current connection has been closed (by the sub-reports) and in that case, create a new session and return a new connection.

--HG--
extra : rebase_source : 29d2c3890b58671b6d968c9c7df5a96ace780e8c