fixes issue 35106: The prompt text is not shown on the multiple selector grids
The selector component was not configuring the showHover property. Now it is configured in the selector fields by checking the UIDefinition of each one. The UIDefinition has a new showHover method which defines if a reference supports hover. It returns true by default but it can be overriden by any subclass in case hover texts are not desired, like in the case of PasswordUIDefinition.

Together with this, the prompt for the checkbox fields of the multi-selector grid has been enabled.