Related to issue 34993: Code review improvements
Removed unnecessary variables and query.
If the invoice line will be deleted always the invoice must update it prepayment
amount. If the invoice line is related to a paid order line, then invoice
prepayment will be updated discounting the prepaid value of the order line, in
other hands the prepayment amount of the invoice will still with its value because
there are not any pay related to the order line.