Fixed issue 35214:Dropdowns are not opening properly when included in a process
The problem is assuming that the selector is placed on a standard view (window).
When a selector is placed on a process definition, the selector raises an error
and no data is showed. It must consider the case where the selector is being
displayed from a process definition view.

This problem has been resolved by take into account when a selector is being
displayed from a process definition. Now, the problem has been fixed.