Fixes issue 35252: General Ledger filter changes with pagination in JE Report
Selected General Ledger filter value changes when pagination is executed in the
Journal Entries report

The General Ledger filter is not kept in the session to make use in the pagination.
For the solution, was saved the General Ledger Filter in the session every time
the search is executed (FIND action). By default, if there is any GL in the
session, the GL field will use the defaulted for the selected org. When paginating
always is used the account schema id stored in the session.