fixes issue 35319: documents were not being attached after send them by e-mail
The CoreAttachImplementation class is deleting the original report file after completing the upload process. When attaching a document using the e-mail pop-up this is done before attaching the document into the e-mail. This explains why the document was not being received on the e-mail.

To avoid this problem, now a copy of the original document is sent to the AttachImplementationManager, this way the original file will never be deleted by an upload process handler.

Together with this, now when attaching a document after printing (not by e-mail) we have to delete the original file because in other case it will be kept on disk.