Fixes issue 34993: Update the invoice prepayment amount with correct value if after create from a paid order line, it updates the line net amount to 0 or it changes from 0 to a another value.
If you create the invoice line from a paid order line, and then the line net amount
changes it value to 0 (by updating qty = 0 or price = 0, for instance), then
previously prepayment was remaining instead of be discarded. So, you may have an
invoice with Total Gross Amount of 0 and a prepayment different than 0.

When you are updating a line with a NEW line net amount = 0 or deleting the line,
it is needed to remove the prepayment of this line from the invoice.

If the OLD line net amount was zero (the prepayment was removed or not taken into
account when creating or updating the line) and it changes
to a non-zero new line net amount value, is necessary take into account the prepaid
order line amount to increase the invoice prepayment amount with its value.

In any other case it is not necessary update the prepayment amount of the invoice.