Fixed issue 35205: Partial indexes functional stuff
The only performance drawback in transforming a normal index to a partial index is in the queries with a where clause filtering by a indexed column when this column is null.
In this case, and only when the ratio of null values is very low, the query will probably execute a seq. scan over the table instead of using the index, thus killing the performance.

To detect this scenario, we have "greped" the PI code for "is*null" string (case insensitve), and filtered the places where the candidate indexes appear. Then, we have individually analize each of the queries to detect whether we could safely transform the index to partial index based on the previous explanation.

The result can be found in https://docs.google.com/spreadsheets/d/1m5GZaaoEkYvyGdyxddtxxb3SeAhvtavfdq53iumOT7s Functional Analysis tab.
Values with Candidate = Y are the ones to be transformed to partial indexes, Candidate = N means to keep the normal index, and TBCN is to create a normal index because it would help in performance.
When there is no query that uses this index inside an "IS NULL" condition, we have automatically set it to partial index as there shouldn't be any risk
For the rest of the candidates there are individual explanations which can be found as a note (in the Candidate column and/or in the places where this index might be used).

For future references, this is the command used to grep the PI code:
grep -n --color=always  -riI  "is[[:space:]]*null" --exclude \*.js --exclude-dir src-wad --exclude-dir build  /home/openbravo/HG/erp/devel/pi | grep -vi case | more

--HG--
extra : transplant_source : y%BAt%17p%A0%F2%11%7D%0B%D0%98B%BE%81%EB%02%9D%AA%A0