Fixed bug 35287: Slow behavior in Financial Account | Accounting tab
Added new index to improve query performance.
In the customer environment the total cost has been reduced from 6581196 to 1310605, and the query time from 18 seconds to 1.2 seconds.

The index is declared as partial index because, although the null ratio is going to be low (around 10% in real instances), it is very unlikely to search by "is null" in this column. So we can have a (small) benefit in declaring as partial index (faster inserts/deletes/updates when the column is null).

--HG--
extra : rebase_source : 70426d5d252bffa6a37897cebdd20de7f3ca8d26