Fixes Issue 35054:It is not possible to pay a negative purchase invoice using credit and selecting "Refund amount to customer"
It is not possible to make use of "Credit To Use"  after processing
Reversed Invoice (Sales / Purchase) or negative AP or AR Invoice if
there are credit payments for business partner used in such invoices.

In Add Payment process definition it not possible to use Credit To Use
whenever selected order/invoices section records sum up expected amount
to a negative value.