fixes bug 35356: print & attach doc logs error using (old) commons jdbc pool
After closing a connection retrieved from the Hibernate pool, it is marked as not usable. It is not possible to check the status of such connections because a HibernateException is thrown.

After printing a report which makes use of the getCountryNumberFormat and applyCountryDateFormat methods of the Utility class, the connection is closed to prevent possible connection leaks. Therefore, when using the Hibernate pool, the connection is marked as not usable causing the HibernateException when checking the status of the connection after printing the report.

To fix the problem now we are handling the HibernateException and initializing the connection properly.

--HG--
extra : rebase_source : 275f9dfbeb4876686a27610b4ca31c1c78afb83f