Fixes issue 24905: Sales Order relation is not updated in Shipment header in some cases.
When a line was deleted was not verified if there was other
invoice lines associated to orders distinct than the header's Sales/Purchase
Order field and it was not updated rightly. Also when inserting a line from the
create lines process, there being two or more lines already inserted related to
different orders, the Sales Order field related with the header was updated
with the last one.

To get the sales order field updated correctly depending on the previous cases,
the following logic was added to the CreateFrom_Shipment_data.xsql and M_InOut
function.

Each time a line is inserted, the shipment's order line field will be updated
with NULL if there are not any line created from an order or if there are more
than one line created from different orders. In the case that lines were created
from an unique sales order then it will be fullfilled with it, otherwise it will
be left empty.

When the line is deleted, the logic explained above needs to be checked inside
a trigger, but it is not possible because the table is in a mutant status in
Oracle, due this reason it is not possible to update the shipment's header at
line's deletion time. As solution to this situation, this field is updated when
the shipment is completed.