Fixes issue 35304: Sales and purchase invoice dimensional reports after up gradation (version pr 16Q 2.4) the number columns are exported as text.
Sales Invoice and Purchase Invoice dimensional reports were not designed considering the
type of the original text field expressions when exported it to xls format.

The net.sf.jasperreports.export.xls.detect.cell.type property was added to
autodetect the cell types and also the pattern "#,##0.00;-#,##0.00" was
included to the fields of number type to rightly format them.