fixes issue 35399: Check readable clients for child properties using DAL WS
Fix for issue #29683 was introducing a security check to avoid showing child properties which belong to a client different from the current client. But this fix was preventing to display some readable objects like master data defined at system level (client 0).

Therefore this security check has been improved and now it is verified if the child property is defined for a client which is readable for the current role.

--HG--
extra : rebase_source : e00845c8475eec27c06ce2907795d3c3b2481545