Fixed bug 32493: Pareto Product Report performance refactor
The Pareto Product report had very important performance issues due to the way data was retrieved from the database.
Multiple (and unnecessary) sequencial scans were executed on high volume tables making the report unusable on real environments.
Besides it has been detected that the information shown was not right: products without stock might appear in the report, the percentages were clearly not right and the ABC category was also wrong because it was based on wrong percentages.

The fix mainly includes a total refactor of the queries, which fixes most of the performance issues.
It also adds data aggregation support based on Valued Stock aggregated data, which should help to keep the report's performance when the time goes by.

It has been tested that, even without aggregated data, the report performs really well on high volume environments. Example: An environment without aggregated data and with 7.7E+6 transactions took only 38 seconds to get the data for an organization with 250E+3 transactions; before it took "days".

List of important changes:
* Deprecated M_GET_PARETO_ABC function and implemented the same logic directly into the SQL query

* Added index on M_TRANSACTION_COST table to the DATEACCT column, which is heavily used as a filter criteria for many reports (included Pareto Product)

* Modified index M_VALUED_STOCK_AGG_DATETO on M_VALUED_STOCK_AGG to take into account AD_ORG_ID and DATETO. This makes the queries to get data from this aggregated table to perform really well.

* Modified ReportParetoProduct_data.xsq select query. This is actually the key to get the performance improvement:
 * Usage of window functions to calculate:
  * the total value per organization's warehouse,
  * the percentage per product and organization's warehouse,
  * the accumulated percentage per product and organization's warehouse.

 * The ABC is calculated looking at the accumulated percentage gotten before instead of calling the M_GET_PARETO_ABC db function.

 * Reduced the usage of functions in select clause because, although individually they are quite fast, they must be executed over multiple records thus making the whole query really slow (example 2ms multiplied by 10E3 records delays the query 20 seconds!):
  * ad_get_org_le_bu is not needed anymore because the legal entity will be always the same for each record in the report. Note that the Pareto's organization combo only shows legal entities or child organizations (which only have a unique legal entity).
  * c_currency_convert_precision is only called when a currency conversion is needed, thus reducing unnecessary overhead.
  * ad_column_identifier calls have been removed. Instead we get static (translated) values.

 * Transactions are now filtered by trxprocessdate when the costing rule was started, so we avoid to compute legacy records.

 * Although not strictly necessary to get a performance improvement, the query has been split in several CTEs, thus making the query more readable and easy to maintain.