Fixes issue 35420:Cannot remove an orphan line in Return to/from vendor/customer
Was not able to remove an orphan line from a return to vendor or a return from
customer document because an incorrect casting to boolean.

When is validating tha Last record is selected inside the Orphan grid lines, the
lastRecord.get("obSelected") object gets as String instead of a Boolean object,
in other case it gets as Boolean.

Now, conditions are rightly validated depending of the class of the object.