Fixes issue 35434: Manual taxes deleted when reactivating a PIT invoice
Now when reactivating a price including taxes invoice, manual taxes will not be deleted, and invoice gross amount will be updated to the sum of tax amounts of the manual taxes instead of zero.
Also, when calculating invoice gross and net amounts in invoice line trigger, manual taxes will be taken into account when getting the sum of tax amounts but not for getting the sum of tax base amounts.