Fixes issue 35302: Updated the node's organization of the product category tree.
When you change the organization of a product category, it is not changed from
the node of the product category tree. As consequence, categories are not shown
in webPOS.

To solve this, when the product category's organization is changed, the node's organization of the
product category tree is updated too.