Fixed issue 35435. Centralize in protected method user and password check
New protected method checkUserPassword() has been added in AuthenticationManager
to centralize all the calls to check the user and password.

The DefaultAuthenticationManager now catches AuthenticationExceptions in case
the implementations of the new method throw it instead of returning null user
id.

checkIfPasswordExpired() method changed to protected() to allow
AuthenticationManager implementations to customize the password expiration
check.