Fixed bug 35442: Explicity set isSOTRX in netting shipments
The netting shipments were created without explicity setting the isSOTRX information. If not set, the database automatically sets it to Y.
However when an EntityEventObserver is executed, this value is null and could raise a NPE.

Solution sets the isSOTRX value exactly the same as the order from which it was created. Right now it will be always Y, but done this way to reuse this method in purchase flows if necessary.

--HG--
extra : transplant_source : v%D2%25%FC%1C%9Dc%80%A3%9D%D9-%406%D7%D8L%21%BA%DA