[login-cleanup] Keep redirect when command is FORCE_NAMED_USER
Do the redirect if the command is FORCE_NAMED_USER. This means that we are doing a login with a user that already has an existing session active having a license that prevents multiple sessions for the same user.
In this case, we perform a redirect directly as at this point the login response has already been handled in the client.