[login-cleanup] Avoid redirection to /security/Menu.html on login
After login the LoginHandler does not redirect to /security/Menu.html anymore. Before this change, the StartPage preference was calculated after that redirection. Now the preference is calculated in the LoginHandler itself.

Together with this, the Preferences API has been increased with the ability to provide the default filters (active, readable clients and readable orgs) to look for the desired preference. This was needed because the retrieval of the StartPage preference in the LoginHandler was being done with the system user context.