Fixes issue 35371: Wrong orders shown in "Create Invoices From Orders" window
Fixed the C_INVOICE_CANDIDATE_V view, changing the condition to not take into account
the sales orders different than 'ON' AND 'OB' AND 'WR' instead of using an incorrect
OR condition as suggested in the proposed solution.

Also were included some performances in the view replacing OR and AND conditions with they
respectively translations to IN and NOT IN clauses.