Fixed issue 35464: Adaptive filtering does not work for Table references.
The problem was that adaptive filtering (filter in client when all data is locally available) did
not work for Table references when filtering by text instead of by selecting a value in the drop
down list. The issue that causes this regression was not taking into account adaptive filtering.

This problem has been fixed by taking into account this case in getDisplayProperty method. Now
getDisplayProperty method manages properly when all the data are in the client (Adaptive filtering).