Fixed issue 35489: Wrong priority between different possible translations
The problem was when you have an especific translation for a static text in a report (registered
in ad_textinterfaces with the filename column pointing to the repport) and there was a general
translation for the same static text. In this situation, the query that returns the possible
translations prioritizes the general translation over the specific translation.

The problem has been fixed by edit the query and prioritizes a especific translation for a static
text in a report instead of the general translation. Besides it have been removed some unnecessary
lines of code.