[pi-application-performance-improvements] quick create and launch refactorized
Before, there were two requests done for the creation of the datasource of the quick launch and the quick create components. There were two more requests to the datasource of each component when they were accessed.

Those requests have been removed by being able to generate the quick launch and quick create components from the menu. So, no datasorurces are needed for this task. Now the quick launch and quick create components are lazily the first time the component is accessed.