Fixes issue 35486: Net Amount not properly rounded in ERP with PostgreSQL
Problem was in following calculation: Net Amount = Gross Amount * (Gross Amount  / (Gross Amount + Tax Amount)).
In case Gross Amount = 357.99 and Tax Amount = 71.598, Net Amount = 357.99 * (357.99 / (357.99 + 71.598)) = 298.325.

It was correct in Oracle as it retrieves 298.325 but it was failing in PostgreSQL as it retrieves 298.3249999999999999988067:
SELECT 357.99 * (357.99  / (357.99 + 71.598)) FROM dual;

If we do multiplication before division, we avoid this problem and both PostgreSQL and Oracle retrieves the correct result (298.325):
SELECT (357.99 * 357.99)  / (357.99 + 71.598) FROM dual;