Fixes issue 35498: Error while posting a Good Receipt related an Order is Void
The error is caused because when post a Goods Receipt using
"Book Using Purchase Order Price" and order is Voided, is being used
the net amount of the line corresponding to the Order instead of the net price.

To solve this problem, consider the proposed solution by modifying the
getDocumentConfirmation method so that it takes into account the net price
of the line instead of its net amount.