[login-perf] Make filtering work properly in quick menus
the getClientPickListData() function must return the whole data available in the combo item. Therefore, this.pickList.data should not be used as it contains the current elements matching with the filtering criteria. This caused that the available data was being lost with every new filtering operation.

To fix this problem we now use a new attribute called entities which will always contain the elements available for the current menu.