Fixed issue 35081: Wrong behaviour when the attribute included in an attribute set is named lot, serno or DateFrom
When the attribute included in an attribute set is named lot, serno or DateFrom
its values are saved in the lot, serno and guaranteedate of the m_attributesetinstance
table, which are reserved for attribute sets defined as Lot, Serial No. and
Expiration date respectively.

To avoid this unliked behaviour, were renamed the Lot, Serial No. and Expiration
date params with more complex identifiers with the suffix "xxxinternalOB" that
identifies from rest of the attributes created from the Application dictionary.