Fixes issue 35499: Currency conversion problem when posting a Goods Receipt.
This problem is caused because when posting a Goods Receipt related to a
Purchase Order and referring product defined as
"Book Using Purchase Order Price", is taking into account the Currency of the
Legal Organization to which Goods Receipt document belongs in exchange
for the Purchase Order.

To correct this error, is takes into account the Currency of the Purchase Order
related with the lines of the Goods Receipt, in the posting process.

--HG--
extra : rebase_source : b51ca532e7b724eb0eaefbc75a09a6b46878edee