Fixes issue 35539: Delivered Sales Order Report is not showing business partner
The problem is that the link between c_order and c_bpartner tables is made by
the c_project table and in cases where the sales order is not related to a
project, the relationship between c_project and c_bpartner is going to return
null, causing it to be impossible to obtain the information of the
business partner.

To solve this problem the union is established between the tables c_order
and c_bdpartner.