[improvements-eac] Calculate static permissions in a static block.
The static queries (Not dependent on the current role) can be executed one time per
instance of EntityAccessChecker class. In order to achieve this, the queries have
been moved to a static block of code.

Before this changeset, the queries in the database filtered the information based on
certain parameters. Now this filter should be applied in the java part instead of in
the database.