Fixes issue 35385: Error is shown when posting a Goods Movement with a 0 cost
When it was a matter of posting a Goods Movements with zero cost, it showed
'Posted error', and in spite of this it was tried to post the document leaving
it in an inconsistent state, when testing the option Unpost an exception was
fired because it was to eliminate the lines When these were not created.

To solve this when it comes to posting a Goods Movements, it is validated that
all lines of this have a cost other than zero or cost null, in case this
condition is not met, the document is changed to Document Disabled and returned
False interrupting the post process since the cost of its lines has zero value.