Fixes issue 35567: Sales Invoice Dimensional Report shows incorrect address
When launch Sales Invoice Dimensional Report and export to XLS, is showing
Invoice Address instead of Shipment Address, because in the query in charge
of obtaining the data, the address information is being retrieved through the
relationship between the C_BPARTNER_LOCATION and C_INVOICE table.

Now to retrieve the correct address, the relationship between
the C_BPARTNER_LOCATION and M_INOUT tables was established.