[login-perf] Removed request to HeartBeatPopupActionHandler
During log-in a request is done to the HeartBeatPopupActionHandler in order to determine if the Heartbeat or any of the registration pop-ups must be displayed. This resquest now is merged with the rest of existing session dynamic resources. The HeartBeatPopUpComponent resource now generates dynamically the content of the new OB.Application.showHeartBeatOrRegistration function which will display the pop-up that applies on each case.