Fixed issue 35536: Deactivated summary function in computed columns.
The summary functions didn't work as expected in computed columns.

As computed columns allow to be defined with any arbitrary sql statement,
the risk of underperformance when included in a summary function is much
bigger.

Because of this, to avoid this problem, it has been resolved by blocking
the use of the summary functions in computed columns.