Fixes issue 35270: Wrong Sales Order reservation with a voided partial shipment
When closing a Sales Order with a reservation that contains a voided partial shipment
The reservation created in the shipment voiding process was keeping associated
to the Sales Order. Also in the Stock Reservation window the Manage Stock button
is show in a reservation with close status.

Now the reservation stock created in the voiding process of the Shipment is deleted
in the close process of the order, also the reservation stock qty = releasedqty
if the order line orderedqty is greater than reservation releasedqty

The display logic of the Manage Stock button now check the selected reservation
status.

--HG--
extra : rebase_source : 2ede8f10ba01115870c99e50ea45d02b19da73cb