Fixed issue 35609: Terminals will detect that terminal authentication preference have been disabled.
If one terminal which is linked detect that the terminal authentication have been disabled, then information related to terminal authentication stored in the localstorage is removed. Apart from that, the process to link terminals has been evolved. Now it detects when a device which is already linked with an specific terminal is trying to do the same link. In this case no error will be raised.

--HG--
extra : rebase_source : 2bdef4678ea84bf3fabbe0ee3226fb73058cfb17