Fixes issue 35610: Unhomogeneous behavior on process used to create sales invoice
Having a Sales Order document with "Order Reference" value, when sales
invoice is created using a process which executes "C_INVOICE_CREATE" function,
extra line with only description information is added in sales invoice document.

As executing "Create Invoices from Orders" process, Order Reference information
is stored on invoice's header, it is not needed to create a line just to log this
information on Description field.

To get an homogeneous behavior, this extra line it is not created.

Were removed from the C_INVOICE_CREATE function all variables and code fragments
used to get this reference and create this extra line with it, as they will be
unused with this solution.