Fixes issue 28073: Wrong Purchase Invoice created by using "Create Lines From"
Purchase Invoice wasn't properly created by using "Create Lines From" process if
a goods receipt has been voided.

When the process was run selecting an order with a goods receipt voided, joins
between tables were adding lines related to the reverse goods receipt because the
goods receipt status wasn't taken into account.

To avoid it, was added to the query the condition that M_InOut.docstatus is
different than void ('VO') to not consider this lines of cancelled receipts.