Fixed issue 35545: No tree view on a sub-tab in some circumstances
The problem was found in an environment with the configured manufacturing
plan and the tree view in the Lines tab of the manufacturing plan. If you
have the tree view on the subtab and you want to generate the data in this
subtab you must start a process (having selected the header record and
started the Production Plan process), no data will appear on the subtab.

The problem was that the doRefreshContents method did not working as
expected with a tree view. When the this.treeGrid.fetchData command is
executed, a call to the datasource is not being made as expected.This is
occurred because SmartClient determines that the call to the datasource is
not necessary (no difference in the criteria is detected).

In order to fix this problem invalidateCache() method is used in order to
force to do a datasource call keeping the current criteria.