Related with issue 35536: summary functions are not allowed in computed columns
The problem found in the code review was if having at least one computed column as summary
the grid status was persisted or the view was saved, after the fix the columns still
appears with its summary. This happened because the original fix did not take this scenario
into account. It means summary functions applied to computed columns in older versions are
not managed properly.

The problem has been resolved by take into account when a computed column had a summary
function defined in older versions. When take place this situation the summary function is
not displayed anymore. Besides, it is not possible to define a new summary function in a
computed column.