[login-perf] Removed removeInaccessibleEntities method
The logic to show the link to an accessible window in field title is not complete. There are cases where the removeInaccessibleEntities method is not enough to decide whether a role can navigate to a specific window, like when there exists navigation rules.

Besides, as this method is affecting the performance in terms of the generation time of the ApplicationDynamic component, the method has been removed.